Move Permalink and related components out of Menu
