+++ components/index.tsx
@@ -18,4 +18,5 @@ export * from "./layout/InlineCallToAction"
 export * from "./layout/Link"
 export * from "./layout/Markdown"
 export * from "./layout/Menu"
+export * from "./layout/Permalink"
 export * from "./layout/Ref"
